Rename sparkplug-util -> sparkplug/util.
