+++ lib/index.js
@@ -9,7 +9,7 @@ export * from "./debug.js";
 export * from "./deps.js";
 export * as secrets from "./secrets.js";
 export * from "./service-client.js";
-export * from "./sparkplug-util.js";
+export * from "./sparkplug/util.js";
 export * from "./util.js";
 export * as UUIDs from "./uuids.js";
 export * from "./webapi.js";
