+++ lib/service/auth.js
@@ -6,7 +6,7 @@
 
 import semver from "semver";
 
-import { Address } from "../sparkplug-util.js";
+import { Address } from "../sparkplug/util.js";
 import { App, Service, Null as Null_UUID } from "../uuids.js";
 
 import ServiceInterface from "./service-interface.js";
