+++ lib/sparkplug/basic-node.js
@@ -2,7 +2,7 @@ import { EventEmitter } from "node:events";
 
 import { Debug } from "../debug.js";
 import { SpB } from "../deps.js";
-import { Address, Topic } from "../sparkplug-util.js";
+import { Address, Topic } from "./util.js";
 
 export class BasicSparkplugNode extends EventEmitter {
     constructor (opts) {
