+++ lib/sparkplug/util.js
