+++ types/index.d.ts
@@ -10,7 +10,7 @@ import {GotFetchResponse} from "got-fetch/out/lib/response";
 export * from "./db";
 export * from "./debug";
 export * from "./service-client"
-export * from "./sparkplug-util"
+export * from "./sparkplug/util"
 export * from "./webapi"
 
 export * from "./service/auth";
